Work-around for empty JSON object when hendelsesliste is empty.
Skatteetaten's hendelsesliste responds with HTTP 200 OK and the body: `{}` (empty JSON object)
when there are no more hendelser to return. The empty JSON object is not something that can be
de-serialized with Avro to the Hendelsesliste class without changing the underlying schema (or using a different writer schema).

Also, if the response contains hendelser, they are given by the JSON key `hendelse` instead of `hendelser`.
This affects class generation, as it would result in the method getHendelse() instead of the more intuitive getHendelser().

We therefore provide an alternative writer schema below, which uses the `hendelse` key,
and an ugly work-around to short-circuit the empty JSON object case.